+++ src/pymbus/telegrams/records.py
@@ -1,9 +1,6 @@
-from collections.abc import Iterable
+"""M-Bus Telegram Data Record module."""
 
-from pymbus.telegrams.base import (
-    TelegramByteType,
-    TelegramContainer,
-)
+from pymbus.telegrams.base import TelegramByteIterableType, TelegramContainer
 from pymbus.telegrams.blocks import (
     DataInformationBlock as DIB,
 )
@@ -12,16 +9,10 @@ from pymbus.telegrams.blocks import (
 )
 
 
-class TelegramRecord(TelegramContainer):
-    """Base Telegram Record class"""
+class DataRecordHeader(TelegramContainer):
+    """The "Data Record Header" (DRH) class.
 
-
-class DataRecord(TelegramRecord):
-    """The "Data Record" (DR) class.
-
-    Typically encountered as Data Record Header (DRH).
-
-    The structure of the DR(H):
+    The structure of the DRH:
     -----------------
     |   DIB  |  VIB |
     -----------------
@@ -30,22 +21,61 @@ class DataRecord(TelegramRecord):
     VIB = Value Information Block.
     """
 
-    def __init__(
-        self, ibytes: None | Iterable[TelegramByteType] = None
-    ) -> None:
-        container = list(TelegramContainer(ibytes=ibytes))
+    def __init__(self, ibytes: None | TelegramByteIterableType = None) -> None:
+        it = iter(ibytes)  # type: ignore [arg-type]
 
-        dib = DIB(ibytes=container)
-        vib = VIB(ibytes=container[len(dib) :])
+        dib = DIB(ibytes=it)
+        vib = VIB(ibytes=it)
 
-        super().__init__(ibytes=container)
+        fields = list(dib) + list(vib)
+        super().__init__(ibytes=fields)
         self._dib = dib
         self._vib = vib
 
     @property
     def dib(self) -> DIB:
+        """Return DI block."""
+
         return self._dib
 
     @property
     def vib(self) -> VIB:
+        """Return VI block."""
+
         return self._vib
+
+
+class DataRecord(TelegramContainer):
+    """The "Data Record" (DR) class.
+
+    The structure of the DR:
+    ----------------
+    |   DRH | data |
+    ----------------
+
+    DRH = Data Record Header.
+    data = data associated with the DRH.
+    """
+
+    def __init__(self, ibytes: None | TelegramByteIterableType = None) -> None:
+        it = iter(ibytes)  # type: ignore [arg-type]
+
+        drh = DataRecordHeader(ibytes=it)
+        data = TelegramContainer(ibytes=it)
+
+        fields = list(drh) + list(data)
+        super().__init__(ibytes=fields)
+        self._drh = drh
+        self._data = data
+
+    @property
+    def drh(self) -> DataRecordHeader:
+        """Return DataRecordHeader."""
+
+        return self._drh
+
+    @property
+    def data(self) -> TelegramContainer:
+        """Return data fields."""
+
+        return self._data